[ISSUE-70] Record Class
- Record Class의 getId()의 키 추출이 하드코딩에서 로직으로 바뀜
- int를 반환해야하는데 컬렉션에 java.lang.Long 타입인 문제 발견
- resultMap을 써서 List<Map>타입으로 반환하면서 문제인 것을 찾음
- VO 클래스 만들고 문제 해결
- Record Class를 위한 Mapper 따로 생성

+++ src/main/java/com/murmuler/organicstack/controller/SearchController.java
@@ -50,9 +50,9 @@ public class SearchController {
     @RequestMapping(value = "", method = RequestMethod.GET)
     public ModelAndView searchRoom() {
         ModelAndView mav = new ModelAndView();
-        List<Map<Long, String>> options = new ArrayList<>();
-        for (long key : optionRecord.keySet()) {
-            Map<Long, String> ops = new HashMap<>();
+        List<Map<Integer, String>> options = new ArrayList<>();
+        for (int key : optionRecord.keySet()) {
+            Map<Integer, String> ops = new HashMap<>();
             ops.put(key, optionRecord.get(key));
             options.add(ops);
         }
